refactor(HomeEventHandler): 홈 화면 내 셀렉트박스 모듈화 및 중복 로직 제거
