+++ src/eventHandler/HomeEventHandler.js
@@ -1,5 +1,10 @@
+import RestaurantComponent from '../components/Restaurant';
+
 class HomeEventHandler {
-  constructor() {
+  restaurantList;
+
+  constructor(restaurantList) {
+    this.restaurantList = restaurantList;
     this.setEvent();
   }
 
@@ -9,6 +14,24 @@ class HomeEventHandler {
 
   handleOpenModal() {
     document.querySelector('.modal').classList.add('modal--open');
+  }
+
+  handleFilter() {
+    const $categoryFilter = document.getElementById('category-filter');
+    const $sortingFilter = document.getElementById('sorting-filter');
+    const $restaurantList = document.querySelector('.restaurant-list');
+
+    const categoryOptions = $categoryFilter.options;
+    const category = categoryOptions[categoryOptions.selectedIndex].text;
+    const sortingOptions = $sortingFilter.options;
+    const sortingCondition = sortingOptions[sortingOptions.selectedIndex].text;
+
+    this.restaurantList.filterByCategory(category);
+    const sortedList = this.restaurantList.getSortedByCondition(sortingCondition);
+    $restaurantList.replaceChildren();
+    sortedList.forEach(element => {
+      $restaurantList.innerHTML += RestaurantComponent(element.information);
+    });
   }
 }
